LayoutTests/media/video-src-set-expected.txt

Test that setting src attribute triggers load
EVENT(loadstart)
EXPECTED (stripExtension(relativeURL(video.currentSrc)) == 'content/test') OK
END OF TEST
